Adding support for arch linux manjaro os
The `install-redpesk-localbuilder.sh` script is not handling the
support of arch linx (manjaro) system
This fix attempt to solve it by adding the proper command line when
installing packages (add a condition in the switch condition)
Furthermore it adds a check to see if the VERSION_ID variable is set
after greping the `/etc/os-release` file. If not, it searches in
`/etc/lsb-release` file. However, to avoid the script exiting after
the first grep failure, we need to remove the `set -e` command at
the beginning of the script.

Resolve: #3

Signed-off-by: antoine.treboutte <antoine.treboutte@iot.bzh>