Now able to fetch an image by fingerprint
We used to only be able to specify an image by name, this commit enables
using the fingerprint as well.

If a match fails for the name, then the script looks for a fingerprint
before reporting a failure. This allows to install container candidates
from other remotes when they do not have any alias defined.