[4650] Allow user to not install recommended dependencies
Modification of the 'install-redpesk-sdk.sh' script in order to allow
the user to not install the recommended packages.

For the moment it will be mainly used for a light installation of
`rp-cli`: without `git`, `git-lfs` and `less`.

The flag to pass to the script to not install the recommended package is
`--no-recommends` or `-n`.

Signed-off-by: Armand Bénéteau <armand.beneteau@iot.bzh>